% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ww_data.R
\name{add_correct_lod}
\alias{add_correct_lod}
\title{Add the correct LOD using external data on LOQ for each gene at each site
and time point}
\usage{
add_correct_lod(ww_data, path_to_lod_vals = NULL)
}
\arguments{
\item{ww_data}{Data.frame of wastewater data from an individual location}

\item{path_to_lod_vals}{Character string indicating the file path to the LOQ
data}
}
\value{
Data.frame containing updated \code{ww_data} where the LOD column has
now been filled in with the geometric mean of the LOQ across all genes
measured on that date and in that site, if the data is flagged as being
below the LOD.
}
\description{
Add the correct LOD using external data on LOQ for each gene at each site
and time point
}
