% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{add_gam_metadata}
\alias{add_gam_metadata}
\title{Attach scaling and date information to a fitted GAM}
\usage{
add_gam_metadata(gam_fit, prepped)
}
\arguments{
\item{gam_fit}{fitted GAM}

\item{prepped}{output of \code{prep_gam_data()}}
}
\value{
GAM object with \code{standardized}, \code{scaling_params} and
\code{min_forecast_date} elements
}
\description{
Attach scaling and date information to a fitted GAM
}
