% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rolling_sum_to_inc.R
\name{convert_rolling_sum_to_inc}
\alias{convert_rolling_sum_to_inc}
\title{Convert rolling sums to incidence data}
\usage{
convert_rolling_sum_to_inc(
  rolling_sums,
  k = 7,
  initial_values = NULL,
  return_same_length = TRUE
)
}
\arguments{
\item{rolling_sums}{A numeric vector of k-day rolling sums}

\item{k}{Integer indicating the number of days that the rolling sums are
computed over. Default is 7.}

\item{initial_values}{A numeric vector of the first k-1 incident values
that contribute to the first rolling sum. If NULL, assumes zeros for the
initial 6 days.}

\item{return_same_length}{Boolean indicating whether or not to return the
data of the same length as the rolling sum data or whether to append the
initial values on.}
}
\value{
A numeric vector of incident values.
}
\description{
This function converts k-day rolling sums of count data back to daily
incidence values. It uses the difference between successive rolling sums
to recover the incident values, handling the initial period appropriately.
}
\examples{
# Example with simple data
y <- sample(5, 20, replace = TRUE)
rolling_sums <- zoo::rollsum(y, k = 7, na.pad = FALSE)
initial_vals <- y[1:6]
daily_incidence <- convert_rolling_sum_to_inc(rolling_sums,
  k = 7,
  initial_vals
)
daily_incidence
}
