% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_samples.R
\name{draws_for_scoring}
\alias{draws_for_scoring}
\title{Preparing draws for scoring, including choosing quantiles or samples}
\usage{
draws_for_scoring(draws, forecast_date, offset = 1, quantiles = FALSE)
}
\arguments{
\item{draws}{Dataframe of model draws with observed data to score}

\item{forecast_date}{Forecast data}

\item{offset}{Offset to use when transforming forecasts}

\item{quantiles}{Converts to quantiles if TRUE, otherwise leaves as samples}
}
\value{
a dataframe with either samples/quantiles, observed data and forecasts on the log scale
}
\description{
Preparing draws for scoring, including choosing quantiles or samples
}
