% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{exploratory_plot_ww_vs_scores}
\alias{exploratory_plot_ww_vs_scores}
\title{Exploratory plots of rWIS vs wastewater characteristics}
\usage{
exploratory_plot_ww_vs_scores(
  scores,
  ww_metadata,
  plot_type = "continuous",
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs")
)
}
\arguments{
\item{scores}{dataframe of scores by location, forecast date, model, and
target date}

\item{ww_metadata}{Dataframe of wastewater metadata}

\item{plot_type}{Whether you want wastewater characteristics to be
"continuous" or discrete, default is continuous which plots a smooth}

\item{fig_file_name}{Name of figure to save, default is NULL}

\item{fig_file_dir}{FP to save figure}
}
\value{
patchwork object
}
\description{
Exploratory plots of rWIS vs wastewater characteristics
}
