% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_arima.R
\name{fit_arima}
\alias{fit_arima}
\title{Fit ARIMA baseline model}
\usage{
fit_arima(
  hosp_data_for_fit,
  hosp_data_for_eval,
  forecast_date,
  data_right_trunc,
  include_ww,
  model,
  prediction_intervals = c(0.5, 0.9),
  forecast_horizon = 28
)
}
\arguments{
\item{hosp_data_for_fit}{Dataframe of data for a specific location and
forecast date to fit to.}

\item{hosp_data_for_eval}{Dataframe of data for a specific location and
forecast date to evaluate against.}

\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{data_right_trunc}{Boolean indicating whether data is form initial
reports (right-truncated) or updated.}

\item{include_ww}{Boolean indicating whether wastewater is included.}

\item{model}{Character string indicating the name of the model.}

\item{prediction_intervals}{Numeric between 0 and 1 indicating the symmetric
prediction intervals, default is the 50th and 90th so \code{c(0.5, 0.9)}.}

\item{forecast_horizon}{Integer indicating the forecast horizon}
}
\value{
\code{forecast_df} Data.frame of forecasts results in wide format
}
\description{
Fit ARIMA baseline model
}
