% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit the GLM model}
\usage{
fit_glm(scores_to_model, standardize = TRUE)
}
\arguments{
\item{scores_to_model}{wide table of scores with wastewater metadata}

\item{standardize}{logical, whether to z-score covariates before fitting}
}
\value{
GLM object (with scaling attributes if standardize = TRUE)
}
\description{
Fit the GLM model
}
