% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{format_baseline_forecasts}
\alias{format_baseline_forecasts}
\title{Format the baseline forecasts to match the outputs from
wwinference}
\usage{
format_baseline_forecasts(
  baseline_forecasts,
  quantiles_to_save,
  offset = 1,
  fp_data = "output"
)
}
\arguments{
\item{baseline_forecasts}{Data.frame of quantiled baseline forecasts}

\item{quantiles_to_save}{Vector of quantiles to save}

\item{offset}{Offset to use when transforming forecasts}

\item{fp_data}{Character string indicating the high level file path to
save the data.}
}
\value{
Data.frame with the baseline forecasts formatted for scoringutils
}
\description{
Format the baseline forecasts to match the outputs from
wwinference
}
