% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{gam_var_to_original_scale}
\alias{gam_var_to_original_scale}
\title{Convert a GAM covariate back to its original scale}
\usage{
gam_var_to_original_scale(gam_fit, var, x)
}
\arguments{
\item{gam_fit}{GAM object returned by \code{fit_gam()} or \code{fit_gam_long()}}

\item{var}{Name of the covariate}

\item{x}{Values of the covariate on the scale used in the fit}
}
\value{
Values on the original (unstandardized) scale
}
\description{
Convert a GAM covariate back to its original scale
}
