% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_bar_chart_overall_scores}
\alias{get_bar_chart_overall_scores}
\title{Get bar chart of the overall scores}
\usage{
get_bar_chart_overall_scores(scores)
}
\arguments{
\item{scores}{Data.frame of scores from across locations and forecast dates}
}
\value{
ggplot object
}
\description{
Get bar chart of the overall scores
}
