% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_bar_chart_scores_by_loc}
\alias{get_bar_chart_scores_by_loc}
\title{Get bar chart of WIS by location and forecast date}
\usage{
get_bar_chart_scores_by_loc(scores, n_locations = 3)
}
\arguments{
\item{scores}{Data.frame of scores from across locations and forecast dates}

\item{n_locations}{Integer indicating number of locations to plot. Default is
3. If NULL, all locations are plotted.}
}
\value{
ggplot object
}
\description{
Get bar chart of WIS by location and forecast date
}
