% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{get_gam_avg_rel_wis}
\alias{get_gam_avg_rel_wis}
\title{Average relative WIS implied by a GAM}
\usage{
get_gam_avg_rel_wis(gam_fit)
}
\arguments{
\item{gam_fit}{GAM object returned by \code{fit_gam()} or \code{fit_gam_long()}}
}
\value{
exp of the intercept (offset model) or of the include_ww
coefficient (long-format model)
}
\description{
Average relative WIS implied by a GAM
}
