% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{get_glm_coef_table}
\alias{get_glm_coef_table}
\title{Table of exponentiated GLM coefficients}
\usage{
get_glm_coef_table(glm_fit)
}
\arguments{
\item{glm_fit}{GLM object returned by \code{fit_glm()}}
}
\value{
gt table
}
\description{
Table of exponentiated GLM coefficients
}
