% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hosp_data.R
\name{get_hosp_as_of_forecast_date}
\alias{get_hosp_as_of_forecast_date}
\title{Get hospital admissions data as of the forecast date}
\usage{
get_hosp_as_of_forecast_date(
  forecast_date,
  location_name,
  location_abbr,
  forecast_horizon,
  filepath_name = file.path("input", "data", "hosp", "vintages"),
  hosp_data_url =
    "https://raw.githubusercontent.com/robert-koch-institut/COVID-19-Hospitalisierungen_in_Deutschland/refs/heads/main/Aktuell_Deutschland_adjustierte-COVID-19-Hospitalisierungen.csv"
)
}
\arguments{
\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{location_name}{Character string indicating long form name of state.}

\item{location_abbr}{Character string indicating abbreviation of state}

\item{forecast_horizon}{Forecast horizon. Default is \code{28}.}

\item{filepath_name}{Character string indicating file path to save or
extract hospital admissions vintage}

\item{hosp_data_url}{Character string of URL to hospital admissions data}
}
\value{
Data.frame of hospital admissions as of the forecast date
}
\description{
Get hospital admissions data as of the forecast date
}
