% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hosp_data.R
\name{get_hosp_for_fit}
\alias{get_hosp_for_fit}
\title{Filter hospital admissions data for fitting}
\usage{
get_hosp_for_fit(
  hosp_data_eval,
  location_name,
  location_abbr,
  forecast_date,
  forecast_horizon,
  hosp_data_real_time,
  calibration_period = 100,
  lag = 3
)
}
\arguments{
\item{hosp_data_eval}{hospital admissions data for evaluation step}

\item{location_name}{Character string indicating long form name of state.}

\item{location_abbr}{Character string indicating abbreviation of state}

\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{forecast_horizon}{Forecast horizon. Default is \code{28}.}

\item{hosp_data_real_time}{Boolean indicating whether to use the hospital
admissions as of the forecast date or just truncate the evaluation data.}

\item{calibration_period}{Integer indicating the number of days of
hospital admissions calibration data to extract. Default is \code{100}.
If NULL, we use all the data.}

\item{lag}{Integer indicating the number of days from the forecast date of
the latest hospital admission. Default is \code{3}.}
}
\description{
Filter hospital admissions data for fitting
}
