% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hosp_data.R
\name{get_initial_values}
\alias{get_initial_values}
\title{Get initial values of daily admissions}
\usage{
get_initial_values(
  start_date_RKI_data = "2023-01-02",
  deconvolved_data_url =
    "https://raw.githubusercontent.com/KITmetricslab/hospitalization-nowcast-hub/refs/heads/main/data-truth/COVID-19/COVID-19_hospitalizations_preprocessed.csv",
  filepath_name = file.path("input", "data", "hosp")
)
}
\arguments{
\item{start_date_RKI_data}{Character string indicating the date that the
RKI hospitalization data begins, in YYYY-MM-DD format.}

\item{deconvolved_data_url}{Character string of the url for the reporting
triangle from the German Nowcast Hub. Default is from the data at
https://raw.githubusercontent.com/KITmetricslab/hospitalization-nowcast-hub/refs/heads/main/data-truth/COVID-19/COVID-19_hospitalizations_preprocessed.csv #nolint}

\item{filepath_name}{Name of directory to save the raw input wastewater data.}
}
\description{
This function uses the reporting triangle from the German
Nowcast Hub to get the daily counts by reference date from the RKI data
which is provided as 7-day sums. It does so my summing over all of the
delays at each reference date.
}
