% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_draws_w_data.R
\name{get_model_draws_w_data}
\alias{get_model_draws_w_data}
\title{Combining model draws with data for scoring}
\usage{
get_model_draws_w_data(
  fit_obj_wwinference,
  model_output = c("ww", "hosp"),
  include_ww = c("TRUE", "FALSE"),
  model = c("wwinference", "baseline"),
  forecast_date,
  location,
  eval_data
)
}
\arguments{
\item{fit_obj_wwinference}{wwinference_fit object returned when wwinference::wwinference() is run}

\item{model_output}{Type of model output to extract, either "ww" or "hosp"}

\item{include_ww}{Whether wastewater data was included in the model, either "TRUE" or "FALSE"}

\item{model}{Model type, either "wwinference" or "baseline"}

\item{forecast_date}{Forecast date}

\item{location}{Location name}

\item{eval_data}{Dataframe of observed data to compare against model output}
}
\value{
a dataframe containing model draws and observed data for scoring
}
\description{
Combining model draws with data for scoring
}
