% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_plot_draws_w_calib_data}
\alias{get_plot_draws_w_calib_data}
\title{Get a plot of the draws with both calibration and evaluation data}
\usage{
get_plot_draws_w_calib_data(draws_w_data, full_fp)
}
\arguments{
\item{draws_w_data}{Data.frame of draws with data}

\item{full_fp}{Directory to save}
}
\value{
ggplot object theme element_text
}
\description{
Get a plot of the draws with both calibration and evaluation data
}
