% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_plot_model_comparison}
\alias{get_plot_model_comparison}
\title{Get a plot comparing models for a single location across multiple forecast
dates}
\usage{
get_plot_model_comparison(
  quantiles_to_score,
  hosp_data_long,
  forecast_horizon_to_plot = 28,
  historical_data_to_plot = 90,
  scale_selected = "natural",
  facet_models = FALSE,
  fig_fp
)
}
\arguments{
\item{quantiles_to_score}{Dataframe for a single location across multiple
forecast dates}

\item{hosp_data_long}{Dataframe of hospital admissions data for single
location}

\item{forecast_horizon_to_plot}{Integer indicating number of days of horizon
to plot}

\item{historical_data_to_plot}{Integer indicating number of days into the
past to plot}

\item{scale_selected}{Character string indicating which scale to plot,
default is \code{"natural"}}

\item{facet_models}{Boolean indicating whether to facet the outputs, default
is FALSE.}

\item{fig_fp}{Character string indicating file path to save figure.}
}
\value{
ggplot object
}
\description{
Get a plot comparing models for a single location across multiple forecast
dates
}
