% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_plot_scores_by_loc}
\alias{get_plot_scores_by_loc}
\title{Get bar chart of the scores by forecast date}
\usage{
get_plot_scores_by_loc(scores)
}
\arguments{
\item{scores}{Data.frame of scores from across locations and forecast dates}
}
\value{
ggplot object
}
\description{
Get bar chart of the scores by forecast date
}
