% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_scatterplot_scores}
\alias{get_scatterplot_scores}
\title{Get scatterplot of scores by forecast date and location}
\usage{
get_scatterplot_scores(scores)
}
\arguments{
\item{scores}{Data.frame of scores from across locations and forecast dates}
}
\value{
ggplot object
}
\description{
Get scatterplot of scores by forecast date and location
}
