% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_scatterplot_wis_vs_horizon}
\alias{get_scatterplot_wis_vs_horizon}
\title{Ridge plot of relative WIS by forecast horizon}
\usage{
get_scatterplot_wis_vs_horizon(scores_to_model)
}
\arguments{
\item{scores_to_model}{wide table of scores with wastewater metadata}
}
\value{
ggplot object
}
\description{
Ridge plot of relative WIS by forecast horizon
}
