% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_state_pop_data}
\alias{get_state_pop_data}
\title{Get a table of state names and population sizes}
\usage{
get_state_pop_data(
  hosp_url =
    "https://raw.githubusercontent.com/robert-koch-institut/COVID-19-Hospitalisierungen_in_Deutschland/refs/heads/main/Aktuell_Deutschland_adjustierte-COVID-19-Hospitalisierungen.csv",
  filepath_name = file.path("metadata")
)
}
\arguments{
\item{hosp_url}{Hospital admissions data url}

\item{filepath_name}{Directory to save states and population sizes}
}
\value{
Data.frame of states and population sizes
}
\description{
Get a table of state names and population sizes
}
