% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ww_data.R
\name{get_vintage}
\alias{get_vintage}
\title{Use GitHub raw url and commit history to get the url as of a specific target
date.}
\usage{
get_vintage(raw_url, target_date, github_token = NULL)
}
\arguments{
\item{raw_url}{Character string indicating url of latest data.}

\item{target_date}{As of date}

\item{github_token}{Optional Personal Access Token}
}
\value{
Character string of the url to use to get the data as of a certain
date.
}
\description{
Use GitHub raw url and commit history to get the url as of a specific target
date.
}
