% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ww_data.R
\name{get_ww_data}
\alias{get_ww_data}
\title{Get wastewater data
For now, just pull the latest wastewater data and filter to exclude the
most recent dates}
\usage{
get_ww_data(
  location_name,
  location_abbr,
  forecast_date,
  calibration_period = 100,
  lag = 3
)
}
\arguments{
\item{location_name}{Character string indicating long form name of state.}

\item{location_abbr}{Character string indicating abbreviation of state}

\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{calibration_period}{Integer indicating the number of days of
hospital admissions calibration data to extract. Default is \code{100}.}

\item{lag}{Integer indicating the number of days from the forecast date of
the latest hospital admission. Default is \code{3}.}
}
\description{
Get wastewater data
For now, just pull the latest wastewater data and filter to exclude the
most recent dates
}
