% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ww_data.R
\name{get_ww_for_eval}
\alias{get_ww_for_eval}
\title{Get wastewater data
For now, just pull the latest wastewater data and filter to exclude the
most recent dates}
\usage{
get_ww_for_eval(
  location_name,
  location_abbr,
  forecast_date,
  path_to_lod_vals,
  forecast_horizon = 28,
  filepath_name = file.path("input", "data", "ww")
)
}
\arguments{
\item{location_name}{Character string indicating long form name of state.}

\item{location_abbr}{Character string indicating abbreviation of state}

\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{path_to_lod_vals}{Character string indicating the file path to the LOQ
data}

\item{forecast_horizon}{Forecast horizon. Default is \code{28}.}

\item{filepath_name}{Name of directory to save the raw input wastewater data.}
}
\description{
Get wastewater data
For now, just pull the latest wastewater data and filter to exclude the
most recent dates
}
