% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ww_data.R
\name{get_ww_for_fit}
\alias{get_ww_for_fit}
\title{Filter wastewater data for fitting}
\usage{
get_ww_for_fit(ww_data_eval, forecast_date, calibration_period = 100, lag = 3)
}
\arguments{
\item{ww_data_eval}{wastewater data for evaluation step}

\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{calibration_period}{Integer indicating the number of days of
wastewater calibration data to extract. Default is \code{100}.}
}
\description{
Filter wastewater data for fitting
}
