% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_ww_metadata_histograms}
\alias{get_ww_metadata_histograms}
\title{Histograms of wastewater metadata across locations and forecast dates}
\usage{
get_ww_metadata_histograms(ww_metadata)
}
\arguments{
\item{ww_metadata}{Dataframe of wastewater metadata}
}
\value{
patchwork object
}
\description{
Histograms of wastewater metadata across locations and forecast dates
}
