% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{get_yaxis_labels}
\alias{get_yaxis_labels}
\title{Determine y-axis labels for middle location}
\usage{
get_yaxis_labels(loc, locations)
}
\arguments{
\item{loc}{Current location name}

\item{locations}{Vector of all location names}
}
\value{
List with hospital and wastewater y-axis labels
}
\description{
Determine y-axis labels for middle location
}
\keyword{internal}
