% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast_comparison.R
\name{load_hospital_forecasts}
\alias{load_hospital_forecasts}
\title{Load hospital forecast data for multiple dates and locations}
\usage{
load_hospital_forecasts(output_path, forecast_dates, locations)
}
\arguments{
\item{output_path}{Path to the output folder}

\item{forecast_dates}{Character vector of forecast dates}

\item{locations}{Character vector of location names}
}
\value{
List of hospital forecast data frames
}
\description{
Load hospital forecast data for multiple dates and locations
}
