% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{load_ww_forecasts}
\alias{load_ww_forecasts}
\title{Load wastewater forecast data for multiple dates and locations}
\usage{
load_ww_forecasts(output_path, forecast_dates, locations)
}
\arguments{
\item{output_path}{Path to the output folder}

\item{forecast_dates}{Character vector of forecast dates}

\item{locations}{Character vector of location names}
}
\value{
List of wastewater forecast data frames
}
\description{
Load wastewater forecast data for multiple dates and locations
}
