% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{partial_plot}
\alias{partial_plot}
\title{Make a plot of the partial effects}
\usage{
partial_plot(gam_fit)
}
\arguments{
\item{gam_fit}{GAM object}
}
\value{
ggplot
}
\description{
Make a plot of the partial effects
}
