% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{plot_gam_effect_sizes}
\alias{plot_gam_effect_sizes}
\title{Forest plot of covariate effects on relative WIS}
\usage{
plot_gam_effect_sizes(gam_fit, lower_q = 0.1, upper_q = 0.9)
}
\arguments{
\item{gam_fit}{GAM object returned by \code{fit_gam()} or \code{fit_gam_long()}}

\item{lower_q, upper_q}{Quantiles of each covariate to contrast}
}
\value{
ggplot
}
\description{
Forest plot of covariate effects on relative WIS
}
