% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{plot_rel_wis_by_time}
\alias{plot_rel_wis_by_time}
\title{Plot relative WIS over forecast date}
\usage{
plot_rel_wis_by_time(gam_fit, n_grid = 200)
}
\arguments{
\item{gam_fit}{GAM object returned by \code{fit_gam()} or \code{fit_gam_long()}}

\item{n_grid}{Number of grid points}
}
\value{
ggplot
}
\description{
Plot relative WIS over forecast date
}
