% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{plot_scores_fit_gam}
\alias{plot_scores_fit_gam}
\title{Plot fitted against observed WIS}
\usage{
plot_scores_fit_gam(gam_fit, scores_to_model)
}
\arguments{
\item{gam_fit}{GAM object returned by \code{fit_gam()}}

\item{scores_to_model}{wide table of scores the GAM was fit to}
}
\value{
ggplot
}
\description{
Plot fitted against observed WIS
}
