% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{prep_scores_to_model}
\alias{prep_scores_to_model}
\title{Prepare scores to model}
\usage{
prep_scores_to_model(scores_long, ww_metadata)
}
\arguments{
\item{scores_long}{Long form data with scores by model and horizon}

\item{ww_metadata}{Metadata on wastewater at the location and forecast date
level}
}
\value{
Wide dataframe ready to be fit to model
}
\description{
Prepare scores to model
}
