% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{process_ww_data}
\alias{process_ww_data}
\title{Process wastewater forecast data}
\usage{
process_ww_data(
  ww_forecasts,
  ww_later_obs_list,
  forecast_horizon_to_plot,
  min_date_filter,
  max_date_filter,
  min_forecast_date
)
}
\arguments{
\item{ww_forecasts}{Combined wastewater forecasts}

\item{ww_later_obs_list}{List of later observations}

\item{forecast_horizon_to_plot}{Forecast horizon in days}

\item{min_date_filter}{Minimum date for filtering}

\item{max_date_filter}{Maximum date for filtering}

\item{min_forecast_date}{Minimum forecast date}
}
\value{
List with processed wastewater forecasts and observations
}
\description{
Process wastewater forecast data
}
