% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\docType{data}
\name{quantile_metrics}
\alias{quantile_metrics}
\title{Metrics to use for quantile scores}
\format{
An object of class \code{list} of length 8.
}
\usage{
quantile_metrics
}
\description{
Metrics to use for quantile scores
}
\keyword{datasets}
