% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hosp_data.R
\name{reformat_hosp_data}
\alias{reformat_hosp_data}
\title{Reformat the hospital admissions data (English to German, and daily values)}
\usage{
reformat_hosp_data(
  RKI_hosp_adj,
  location_abbr,
  location_name,
  forecast_date,
  forecast_horizon,
  col_for_final_counts = "aktualisierte_7T_Hospitalisierung_Faelle"
)
}
\arguments{
\item{RKI_hosp_adj}{Data.frame of raw data from GitHub}

\item{location_abbr}{Character string indicating abbreviation of state}

\item{location_name}{Character string indicating long form name of state.}

\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{forecast_horizon}{Forecast horizon. Default is \code{28}.}

\item{col_for_final_counts}{Character string indicating the column to use
for the "final"counts at each reference date. Default is
\code{"aktualisierte_7T_Hospitalisierung_Faelle"} which is the updated counts.}
}
\value{
Data.frame with daily and 7d counts
}
\description{
Reformat the hospital admissions data (English to German, and daily values)
}
