% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ww_data.R
\name{reformat_ww_data}
\alias{reformat_ww_data}
\title{Reformat the ww data, selecting only the location we need and the variables
we need.}
\usage{
reformat_ww_data(raw_ww, location_abbr, location_name, log_lod_val = 1)
}
\arguments{
\item{raw_ww}{Data.frame from the RKI GitHub}

\item{location_abbr}{Character string indicating abbreviation of state}

\item{location_name}{Character string indicating long form name of state.}

\item{log_lod_val}{Scalar indicating the value to reset the
limit of detection (LOD) to, to be
removed in future iterations and replace with an LOD value at each
measurement or site.}
}
\value{
Data.frame of cleaned wastewater values from location_abbr.
}
\description{
Reformat the ww data, selecting only the location we need and the variables
we need.
}
