% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_samples.R
\docType{data}
\name{sample_metrics}
\alias{sample_metrics}
\title{Metrics to use for sample scores}
\format{
An object of class \code{list} of length 10.
}
\usage{
sample_metrics
}
\description{
Metrics to use for sample scores
}
\keyword{datasets}
