% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_samples.R
\docType{data}
\name{sample_metrics_quantiles}
\alias{sample_metrics_quantiles}
\title{Metrics to use for quantile scores}
\format{
An object of class \code{list} of length 8.
}
\usage{
sample_metrics_quantiles
}
\description{
Metrics to use for quantile scores
}
\keyword{datasets}
