% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_csv}
\alias{save_csv}
\title{Save a dataframe to a csv and return the path for targets}
\usage{
save_csv(df, filename, path, allow_empty = TRUE)
}
\arguments{
\item{df}{dataframe to save}

\item{filename}{name of dataframe}

\item{path}{directory to save file in}

\item{allow_empty}{Boolean indicating whether to save if df is empty,
Default is \code{TRUE}.}
}
\description{
Save a dataframe to a csv and return the path for targets
}
