% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_samples.R
\name{score_samples}
\alias{score_samples}
\title{Scores samples against observed data}
\usage{
score_samples(draws_for_scoring, metrics = sample_metrics)
}
\arguments{
\item{draws_for_scoring}{Dataframe of samples or quantiles on log scale ready to score}

\item{metrics}{Metrics to use for scoring, Default: sample_metrics}
}
\value{
a dataframe containing scores for each day of forecasting horizon
}
\description{
Scores samples against observed data
}
