% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_su_object.R, R/utils.R
\name{convert_to_su_object}
\alias{convert_to_su_object}
\title{Convert the scores to a scoringutils object}
\usage{
convert_to_su_object(scores_raw)

convert_to_su_object(scores_raw)
}
\arguments{
\item{scores_raw}{Data.frame of scores}

\item{scores_data}{Data.frame from Variant Nowcast Hub GitHub}
}
\value{
scoringutils object

scoringutils object
}
\description{
Convert the scores to a scoringutils object

Convert the scores to a scoringutils object
}
