% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast_comparison.R
\name{create_hospital_plot}
\alias{create_hospital_plot}
\title{Create hospital plot for a single location}
\usage{
create_hospital_plot(
  loc_hosp_forecast,
  loc_hosp_fit,
  loc_hosp_obs,
  loc,
  locations,
  hosp_ylab
)
}
\arguments{
\item{loc_hosp_forecast}{Hospital forecast data for location}

\item{loc_hosp_fit}{Hospital fit data for location}

\item{loc_hosp_obs}{Hospital observations for location}

\item{loc}{Location name}

\item{locations}{All locations (for indexing)}

\item{hosp_ylab}{Y-axis label}
}
\value{
ggplot object
}
\description{
Create hospital plot for a single location
}
