% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast_comparison.R
\name{create_location_plot}
\alias{create_location_plot}
\title{Create combined plot for a single location}
\usage{
create_location_plot(
  loc,
  forecasts_wide,
  fits_wide,
  hosp_obs,
  ww_wide,
  ww_fits_wide,
  ww_obs,
  locations,
  max_ww_sites
)
}
\arguments{
\item{loc}{Location name}

\item{forecasts_wide}{Hospital forecast data}

\item{fits_wide}{Hospital fit data}

\item{hosp_obs}{Hospital observations}

\item{ww_wide}{Wastewater forecast data (can be NULL)}

\item{ww_fits_wide}{Wastewater fit data (can be NULL)}

\item{ww_obs}{Wastewater observations (can be NULL)}

\item{locations}{Vector of all location names}

\item{max_ww_sites}{Maximum WW sites across all locations}
}
\value{
A combined plot for the location
}
\description{
Create combined plot for a single location
}
\keyword{internal}
