% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast_comparison.R
\name{create_ww_plot}
\alias{create_ww_plot}
\title{Create wastewater plot for a single location}
\usage{
create_ww_plot(
  loc_ww_forecast,
  loc_ww_obs,
  ww_ylab,
  n_ww_sites_loc,
  max_ww_sites
)
}
\arguments{
\item{loc_ww_forecast}{Wastewater forecast data for location}

\item{loc_ww_obs}{Wastewater observations for location}

\item{ww_ylab}{Y-axis label}

\item{n_ww_sites_loc}{Number of WW sites for this location}

\item{max_ww_sites}{Maximum WW sites across all locations}
}
\value{
ggplot or patchwork object
}
\description{
Create wastewater plot for a single location
}
