% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{draws_for_scoring}
\alias{draws_for_scoring}
\title{Preparing draws for scoring, including choosing quantiles or samples}
\usage{
draws_for_scoring(
  draws,
  forecast_date,
  offset = 1,
  quantiles = FALSE,
  remove_before_forecast_date = FALSE,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95)
)
}
\arguments{
\item{draws}{Dataframe of model draws with observed data to score}

\item{forecast_date}{Forecast date}

\item{offset}{Offset to use when transforming forecasts}

\item{quantiles}{Converts to quantiles if TRUE, otherwise leaves as samples}

\item{remove_before_forecast_date}{Boolean indicating whether or not
to remove the values before the forecast date, default is FALSE.}

\item{probs}{Numeric vector of quantile probabilities to use
when \code{quantiles = TRUE}}
}
\value{
A forecast_sample or forecast_quantile object (depending on
the quantiles parameter) on the log scale, ready for scoring with
generate_scores(). Returns NULL if draws is NULL.
}
\description{
Preparing draws for scoring, including choosing quantiles or samples
}
