% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_arima.R
\name{fit_arima}
\alias{fit_arima}
\title{Fit ARIMA baseline model}
\usage{
fit_arima(
  hosp_data_for_fit,
  hosp_data_for_eval,
  forecast_date,
  location_name,
  location_abbr,
  data_right_trunc,
  include_ww,
  model,
  forecast_horizon = 28
)
}
\arguments{
\item{hosp_data_for_fit}{Dataframe of data for a specific location and
forecast date to fit to.}

\item{hosp_data_for_eval}{Dataframe of data for a specific location and
forecast date to evaluate against.}

\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{location_name}{Character string indicating long form name of state.}

\item{location_abbr}{Character string indicating abbreviation of state}

\item{data_right_trunc}{Boolean indicating whether data is form initial
reports (right-truncated) or updated.}

\item{include_ww}{Boolean indicating whether wastewater is included.}

\item{model}{Character string indicating the name of the model.}

\item{forecast_horizon}{Integer indicating the forecast horizon}
}
\value{
\code{forecast_df} Data.frame of forecasts results in wide format
}
\description{
Fit ARIMA baseline model
}
