% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit the GAM model}
\usage{
fit_gam(scores_to_model)
}
\arguments{
\item{scores_to_model}{wide table of scores with wastewater metadata}
}
\value{
GAM object
}
\description{
Fit the GAM model
}
