% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit the GLM model}
\usage{
fit_glm(scores_to_model)
}
\arguments{
\item{scores_to_model}{wide table of scores with wastewater metadata}
}
\value{
GLM object
}
\description{
Fit the GLM model
}
