% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wwinference_wrapper.R
\name{fit_wwinference_wrapper}
\alias{fit_wwinference_wrapper}
\title{Wrapper function to fit the wwinference model and generate the
calibration and forecast quantiles}
\usage{
fit_wwinference_wrapper(
  ww_data,
  count_data,
  hosp_data_eval,
  this_forecast_date,
  model_spec,
  fit_opts,
  compiled_model,
  calibration_time = 90,
  forecast_horizon = 28,
  quantiles_to_save = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975),
  ind_filepath = file.path("output"),
  save_draws = FALSE
)
}
\arguments{
\item{ww_data}{Dataframe of wastewater data for fitting}

\item{count_data}{Dataframe of case data for fitting}

\item{hosp_data_eval}{Dataframe of hospital admissions data for evaluation}

\item{this_forecast_date}{Character string indicating forecast date}

\item{model_spec}{List of model specifications}

\item{fit_opts}{List of fitting specifications}

\item{compiled_model}{Compiled model}

\item{calibration_time}{Integer indicating number of days to calibrate the
model}

\item{forecast_horizon}{Integer indicating number of days to forecast out to}

\item{quantiles_to_save}{Vector of numerics indicating the quantiles}

\item{ind_filepath}{Character string of the file path to save the outputs
from each model run}

\item{save_draws}{Boolean indicating whether or not to save the draws,
default is FALSE.}
}
\value{
Data.frame of the quantiles alongside the evaluation data.
}
\description{
Wrapper function to fit the wwinference model and generate the
calibration and forecast quantiles
}
