% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{generate_scores}
\alias{generate_scores}
\title{Scores samples against observed data}
\usage{
generate_scores(
  draws_for_scoring,
  metrics,
  scale_selected = "log",
  save_scores = FALSE,
  fp_data
)
}
\arguments{
\item{draws_for_scoring}{Dataframe of samples/quantiles on log scale}

\item{metrics}{Metrics to use for scoring}

\item{scale_selected}{Character string indicating whether to score
on natural or log scale, default is "log".}

\item{save_scores}{Boolean indicating whether or not to save the scores.}

\item{fp_data}{Character string indicating filepath to save the scores}
}
\value{
a dataframe containing scores for each day of forecasting horizon
}
\description{
Scores samples against observed data
}
