% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hosp_data.R
\name{get_hosp_data}
\alias{get_hosp_data}
\title{Get the hospital admissions data}
\usage{
get_hosp_data(
  location_name,
  location_abbr,
  forecast_date,
  forecast_horizon = 28,
  right_trunc = FALSE,
  filepath_name = file.path("input", "data", "hosp")
)
}
\arguments{
\item{location_name}{Character string indicating long form name of state.}

\item{location_abbr}{Character string indicating abbreviation of state}

\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{right_trunc}{Boolean indicating whether to use the real-time, right
truncated data or the final corrected data. Default is \code{FALSE} indicating
we will just use the corrected data.}

\item{filepath_name}{Name of directory to save the raw input wastewater data.}

\item{calibration_period}{Integer indicating the number of days of
hospital admissions calibration data to extract. Default is \code{100}.}

\item{lag}{Integer indicating the number of days from the forecast date of
the latest hospital admission. Default is \code{3}.}
}
\description{
Get the hospital admissions data
}
