% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hosp_data.R
\name{get_hosp_for_fit}
\alias{get_hosp_for_fit}
\title{Filter hospital admissions data for fitting}
\usage{
get_hosp_for_fit(
  hosp_data_eval,
  forecast_date,
  right_trunc = FALSE,
  calibration_period = 100,
  lag = 3
)
}
\arguments{
\item{hosp_data_eval}{hospital admissions data for evaluation step}

\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{right_trunc}{Boolean indicating whether to use the real-time, right
truncated data or the final corrected data. Default is \code{FALSE} indicating
we will just use the corrected data.}

\item{calibration_period}{Integer indicating the number of days of
hospital admissions calibration data to extract. Default is \code{100}.
If NULL, we use all the data.}

\item{lag}{Integer indicating the number of days from the forecast date of
the latest hospital admission. Default is \code{3}.}
}
\description{
Filter hospital admissions data for fitting
}
