% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{get_plot_effect_by_location}
\alias{get_plot_effect_by_location}
\title{Get plot of location specific multiplicative effect}
\usage{
get_plot_effect_by_location(gam_fit)
}
\arguments{
\item{gam_fit}{long-format GAM fit from \code{fit_gam_long()}}
}
\value{
ggplot object
}
\description{
Get plot of location specific multiplicative effect
}
