% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_plot_scores_by_horizon}
\alias{get_plot_scores_by_horizon}
\title{Get line plot fo scores by horizons}
\usage{
get_plot_scores_by_horizon(scores)
}
\arguments{
\item{scores}{Data.frame of scores from across locations and forecast dates}
}
\value{
ggplot object
}
\description{
Get line plot fo scores by horizons
}
