% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ww_data.R
\name{get_ww_as_of_forecast_date}
\alias{get_ww_as_of_forecast_date}
\title{Use the Git commit history to get the wastewater data available as of
the forecast date. Subsequent function will filter it to the location we
want.}
\usage{
get_ww_as_of_forecast_date(
  forecast_date,
  location_name,
  location_abbr,
  path_to_lod_vals,
  filepath_name = file.path("input", "data", "ww", "vintages"),
  calibration_period = 100,
  ww_data_url =
    "https://raw.githubusercontent.com/robert-koch-institut/Abwassersurveillance_AMELAG/refs/heads/main/amelag_einzelstandorte.tsv"
)
}
\arguments{
\item{forecast_date}{Character string or date indicating the date of
forecast in YYYY-MM-DD}

\item{location_name}{Character string indicating long form name of state.}

\item{location_abbr}{Character string indicating abbreviation of state}

\item{path_to_lod_vals}{Character string indicating the file path to the LOQ
data}

\item{filepath_name}{Name of directory to save the raw input wastewater data.}

\item{calibration_period}{Integer indicating the number of days of
wastewater calibration data to extract. Default is \code{100}.}

\item{ww_data_url}{Character string of the url of the wastewater data (not
the raw version).}
}
\description{
Use the Git commit history to get the wastewater data available as of
the forecast date. Subsequent function will filter it to the location we
want.
}
