% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{load_later_ww_obs}
\alias{load_later_ww_obs}
\title{Load later wastewater observations for forecast validation}
\usage{
load_later_ww_obs(output_path, forecast_dates, locations)
}
\arguments{
\item{output_path}{Path to the output folder}

\item{forecast_dates}{Character vector of forecast dates}

\item{locations}{Character vector of location names}
}
\value{
List of later wastewater observation data frames
}
\description{
Load later wastewater observations for forecast validation
}
