% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{plot_forecast_comparison}
\alias{plot_forecast_comparison}
\title{Compare forecasts across time}
\usage{
plot_forecast_comparison(
  forecasts_w_eval_data,
  hosp_data_long,
  forecast_horizon_to_plot = 28,
  historical_data_to_plot = 90
)
}
\arguments{
\item{forecasts_w_eval_data}{Data.frame of forecasts with quantiles and
evaluation data}

\item{hosp_data_long}{Data.frame of observed data before the first forecast
date through the last}

\item{forecast_horizon_to_plot}{Integer indicating the number of days to
plot after the last forecast date. Default is \code{28}}

\item{historical_data_to_plot}{Integer indicating number of days to plot
before the first forecast date. Default is 90.}
}
\value{
ggplot object
}
\description{
Compare forecasts across time
}
