% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{plot_forecast_comparison}
\alias{plot_forecast_comparison}
\title{Compare forecasts across time}
\usage{
plot_forecast_comparison(forecasts_w_eval_data, forecast_horizon_to_plot = 28)
}
\arguments{
\item{forecasts_w_eval_data}{Data.frame of forecasts with quantiles and
evaluation data}

\item{forecast_horizon_to_plot}{Integer indicating the number of days to
plot.}
}
\value{
ggplot object
}
\description{
Compare forecasts across time
}
