% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{process_hospital_data}
\alias{process_hospital_data}
\title{Process hospital forecast data}
\usage{
process_hospital_data(
  hosp_forecasts,
  forecast_horizon_to_plot,
  historical_data_to_plot,
  scale_selected
)
}
\arguments{
\item{hosp_forecasts}{Combined hospital forecasts data frame}

\item{forecast_horizon_to_plot}{Forecast horizon in days}

\item{historical_data_to_plot}{Historical data period in days}

\item{scale_selected}{Scale to use}
}
\value{
List with processed hospital forecasts and observations
}
\description{
Process hospital forecast data
}
