% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_samples.R
\name{score_samples}
\alias{score_samples}
\title{Scores samples against observed data}
\usage{
score_samples(draws, metrics = sample_metrics, forecast_date, offset = 1)
}
\arguments{
\item{draws}{Dataframe of model draws with observed data to score}

\item{metrics}{Metrics to use for scoring, Default: sample_metrics}

\item{forecast_date}{Forecast date}

\item{offset}{Offset to use when transforming forecasts}
}
\value{
a dataframe containing scores for each day of forecasting horizon
}
\description{
Scores samples against observed data
}
